// src/generated/client/models/TimeEntryTotalDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Timebank
 * Timebank API documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    TimeEntryTotalId,
    TimeEntryTotalIdFromJSON,
    TimeEntryTotalIdFromJSONTyped,
    TimeEntryTotalIdToJSON,
} from './';

/**
 * 
 * @export
 * @interface TimeEntryTotalDto
 */
export interface TimeEntryTotalDto {
    /**
     * 
     * @type {TimeEntryTotalId}
     * @memberof TimeEntryTotalDto
     */
    id?: TimeEntryTotalId;
    /**
     * 
     * @type {number}
     * @memberof TimeEntryTotalDto
     */
    total: number;
    /**
     * 
     * @type {number}
     * @memberof TimeEntryTotalDto
     */
    logged: number;
    /**
     * 
     * @type {number}
     * @memberof TimeEntryTotalDto
     */
    expected: number;
    /**
     * 
     * @type {number}
     * @memberof TimeEntryTotalDto
     */
    internalTime: number;
    /**
     * 
     * @type {number}
     * @memberof TimeEntryTotalDto
     */
    projectTime: number;
}

export function TimeEntryTotalDtoFromJSON(json: any): TimeEntryTotalDto {
    return TimeEntryTotalDtoFromJSONTyped(json, false);
}

export function TimeEntryTotalDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TimeEntryTotalDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, '_id') ? undefined : TimeEntryTotalIdFromJSON(json['_id']),
        'total': json['total'],
        'logged': json['logged'],
        'expected': json['expected'],
        'internalTime': json['internalTime'],
        'projectTime': json['projectTime'],
    };
}

export function TimeEntryTotalDtoToJSON(value?: TimeEntryTotalDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        '_id': TimeEntryTotalIdToJSON(value.id),
        'total': value.total,
        'logged': value.logged,
        'expected': value.expected,
        'internalTime': value.internalTime,
        'projectTime': value.projectTime,
    };
}